+++ NEWS.md
@@ -10,6 +10,9 @@
 in `tokens_compound()` and `tokens_lookup()` (#2390).
 
 
+* Changes to dfm_select() now include new arguments `match_pattern` and `concatenator`, to control how single-word patterns or multi-word patterns
+  and multi-word patterns matched, as well as controlling the concatenator that binds phrases in the pattern.
+
 # quanteda 4.2.0
 
 ## Changes and additions